Extract mobile dialog drag logic
